+++ src/init.rs
@@ -31,8 +31,13 @@ pub fn init_appdir() -> Result<PathBuf> {
 }
 
 pub fn init_reqwest() -> Result<Client> {
+    #[cfg(target_family = "windows")]
+    let user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/138.0.0.0 Safari/537.36";
+    #[cfg(target_family = "unix")]
+    let user_agent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/138.0.0.0 Safari/537.36";
+
     let c = Client::builder()
-        .user_agent(format!("bread-launcher-{}", env!("CARGO_PKG_VERSION")))
+        .user_agent(user_agent)
         .pool_idle_timeout(None)
         .use_rustls_tls()
         .https_only(true)