+++ src/main.rs
@@ -8,11 +8,17 @@ mod utils;
 mod widgets;
 
 mod account;
+mod init;
 mod instance;
 mod logs;
 mod settings;
 
 fn main() {
+    unsafe {
+        #[cfg(debug_assertions)]
+        std::env::set_var("RUST_BACKTRACE", "1");
+    }
+
     if let Err(e) = app::run() {
         log::error!("{e:#?}");
     }