+++ src/settings.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Settings {
-    jvm_ram: usize, // in MB
+    pub jvm_ram: usize, // in MB
 }
 
 impl Default for Settings {